requirements.txt
streamlit
google-generativeai
notion-client
Pillow
